requirements.txt
aniso8601>=6.0.0
Click>=7.0
Flask>=1.0.2
Flask-JWT-Extended>=3.18.1
Flask-SQLAlchemy>=2.3.2
itsdangerous>=1.1.0
Jinja2==2.11.3
MarkupSafe>=1.1.1
passlib>=1.7.1
PyJWT>=1.7.1
pytz>=2019.1
six>=1.12.0
SQLAlchemy>=1.3.2
Werkzeug>=0.15.2
backports.pbkdf2==0.1
Flask-Cors>=3.0.8
python-dotenv>=0.10.3
flask-mongoengine>=0.9.5
flask-script>=2.0.6
redis>=2.10.6
flask_restplus>=0.3.7
